air quiality fix

+++ AI1/Core/Mappers/TableToAirQualityMapper.cs
@@ -10,19 +10,19 @@ namespace AI1.Core.Mappers {
                 list.Add(new AirQuality {
                     //Id = list.Count,
                     DateTime = DateTime.ParseExact($"{dr[0]}, {dr[1]}", "dd.MM.yyyy, HH.mm.ss", CultureInfo.CurrentCulture),
-                    CO_GT = int.Parse((string)dr[2]),
-                    PT08_S1 = int.Parse((string)dr[3]),
-                    NMHC_GT = int.Parse((string)dr[4]),
-                    C6H6_GT = int.Parse((string)dr[5]),
-                    PT08_S2 = int.Parse((string)dr[6]),
-                    NOx_GT = int.Parse((string)dr[7]),
-                    PT08_S3 = int.Parse((string)dr[8]),
-                    NO2_GT = int.Parse((string)dr[9]),
-                    PT08_S4 = int.Parse((string)dr[10]),
-                    PT08_S5 = int.Parse((string)dr[11]),
-                    T = int.Parse((string)dr[12]),
-                    RH = int.Parse((string)dr[13]),
-                    AH = int.Parse((string)dr[14])
+                    CO_GT = double.Parse((string)dr[2]),
+                    PT08_S1 = double.Parse((string)dr[3]),
+                    NMHC_GT = double.Parse((string)dr[4]),
+                    C6H6_GT = double.Parse((string)dr[5]),
+                    PT08_S2 = double.Parse((string)dr[6]),
+                    NOx_GT = double.Parse((string)dr[7]),
+                    PT08_S3 = double.Parse((string)dr[8]),
+                    NO2_GT = double.Parse((string)dr[9]),
+                    PT08_S4 = double.Parse((string)dr[10]),
+                    PT08_S5 = double.Parse((string)dr[11]),
+                    T = double.Parse((string)dr[12]),
+                    RH = double.Parse((string)dr[13]),
+                    AH = double.Parse((string)dr[14])
                 });
             }
             return list;
